requirements.txt
Django==4.2.1
djangorestframework==3.14.0
numpy==1.23.5
opencv-python==4.7.0.72
torch==2.0.1


